Tidy up features and dependencies
Move sync and no_sync to to [sync|no_sync]_editor
Only build the sync/async features on a build flag
Remove tokio as a noline dependency
