+++ examples/std/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-noline = { path = "../../noline", features = ["std", "tokio"] }
+noline = { path = "../../noline", features = ["std", "async", "sync"] }
 heapless = "0.8.0"
 termion = "4.0.0"
 tokio = { version = "1.16.1", features = [
