+++ noline/Cargo.toml
@@ -18,13 +18,14 @@ embedded-io = "0.6.1"
 embedded-io-async = "0.6.1"
 num_enum = { version = "0.7.2", default-features = false }
 nb = { version = "1.0.0", optional = true }
-tokio = { version = "1.16.1", optional = true, features = ["io-util", "sync"] }
 async-std = { version = "1.12.0", optional = true }
 
 
 [features]
 default = []
 std = ["embedded-io/std", "embedded-io-async/std", "async-std"]
+sync = []
+async = []
 embedded = ["nb"]
 alloc = []
