+++ noline/src/error.rs
@@ -8,13 +8,3 @@ pub enum NolineError {
     ReadError(embedded_io::ErrorKind),
     WriteError(embedded_io::ErrorKind),
 }
-
-// impl Error {
-//     pub fn read_error<T>(err: RE) -> Result<T, Self> {
-//         Err(Self::ReadError(err))
-//     }
-
-//     pub fn write_error<T>(err: WE) -> Result<T, Self> {
-//         Err(Self::WriteError(err))
-//     }
-// }
